# 🔴 refactor: use `Bitfaster.Caching` instead of `Libplanet.LruCahceNet`