--- a/src/Libplanet.Store/Libplanet.Store.csproj
+++ b/src/Libplanet.Store/Libplanet.Store.csproj
@@ -5,6 +5,7 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <PackageReference Include="BitFaster.Caching" Version="2.5.2" />
     <PackageReference Include="System.Collections.Immutable" Version="1.*" />
     <PackageReference Include="Bencodex" Version="0.16.0" />
     <PackageReference Include="Caching.dll" Version="1.4.0.1" />