# Context
Benchmark says at all.

# Benchmark
```
BenchmarkDotNet v0.14.0, macOS Sequoia 15.0 (24A335) [Darwin 24.0.0]
Apple M2 Max, 1 CPU, 12 logical and 12 physical cores
.NET SDK 8.0.101
  [Host]     : .NET 8.0.1 (8.0.123.58001), Arm64 RyuJIT AdvSIMD
  DefaultJob : .NET 8.0.1 (8.0.123.58001), Arm64 RyuJIT AdvSIMD
```
## Read
|    Method      |    Mean   |   Error   |   StdDev  |
|:--------------|:---------|:---------|:---------|
| ConcurrentLru  | 5.156 μs  | 0.0171 μs | 0.0133 μs |
| ConcurrentTLru | 5.372 μs  | 0.0161 μs | 0.0150 μs |
| Planetarium.LruCacheNet       | 15.479 μs | 0.0444 μs | 0.0416 μs |

## Write
| Method         | Mean     | Error    | StdDev   |
|----------------|----------|----------|----------|
| ConcurrentLru  | 26.45 μs | 0.186 μs | 0.165 μs |
| ConcurrentTLru | 26.05 μs | 0.056 μs | 0.047 μs |
| Planetarium.LruCacheNet       | 16.63 μs | 0.119 μs | 0.099 μs |